requirements.txt
pygame
pillow
numpy
opencv-python
configparser
